+++ src/components/modules/tools.ts
@@ -1,4 +1,3 @@
-import Paragraph from '../tools/paragraph/dist/bundle';
 import Module from '../__module';
 import * as _ from '../utils';
 import {
@@ -10,9 +9,6 @@ import {
   ToolConstructable,
   ToolSettings
 } from '../../../types';
-import BoldInlineTool from '../inline-tools/inline-tool-bold';
-import ItalicInlineTool from '../inline-tools/inline-tool-italic';
-import LinkInlineTool from '../inline-tools/inline-tool-link';
 import Stub from '../tools/stub';
 
 /**
@@ -382,17 +378,10 @@ export default class Tools extends Module {
 
   /**
    * Returns internal tools
-   * Includes Bold, Italic, Link and Paragraph
+   * Includes Stub
    */
   public get internalTools(): { [toolName: string]: ToolConstructable | ToolSettings } {
     return {
-      bold: { class: BoldInlineTool },
-      italic: { class: ItalicInlineTool },
-      link: { class: LinkInlineTool },
-      paragraph: {
-        class: Paragraph,
-        inlineToolbar: true,
-      },
       stub: { class: Stub },
     };
   }